// src/lib/expenses/infrastructure/types/gets/get-expenses.interface.ts
export interface IGetExpenses {
  id: string;
  description: string;
  amount: number;
  customerId: string;
  firstName: string;
  lastName: string;
  phoneNumber: string;
  email: string;
  userId: string;
  userName: string;
  personId: string;
  expenseType: string;
  createdAt: string;
  updatedAt: string;
}